Omphile

+++ Controllers/UsersController.cs
@@ -40,31 +40,7 @@ namespace Markers_GPS_Coordiantes.Controllers
                 return Unauthorized("You don't have permission to perform this operation.");  //  write better message
             }
             //  END OF SECURITY CHECK
-            int RoleID = Convert.ToInt32(_sessionAccessor.HttpContext.Session.GetInt32("roleID"));
-            //  filter by role id
-            if ((!(RoleID == (int)RoleIDs.CenterManager)) && (!(RoleID == (int)RoleIDs.SuperAdmin)) && (!(RoleID == (int)RoleIDs.Administrator)))
-            {
-                return null;
-            }
-
-            CenterID = Convert.ToInt32(_sessionAccessor.HttpContext.Session.GetInt32("centerID"));
-            UsersID = Convert.ToInt32(_sessionAccessor.HttpContext.Session.GetInt32("usersID"));
-
-            //  make sure the center really exists
-            var center = await _context.VCenter.Where(b => b.CenterId == CenterID).FirstOrDefaultAsync();
-
-            if (center == null)
-            {
-                return NotFound("The center does not exist");
-            }
-
-            //  check if the user is registered in the CenterManager-Center join table
-            var Accounts = await _context.Users.Where(b => b.UsersId == UsersID && b.CenterId == center.CenterId).FirstOrDefaultAsync();
-            if (Accounts == null)
-            {
-                return NotFound("You are not configured as center manager, please consult your system administrator.");
-            }
-
+           
             ViewData["GetMarkersDetails"] = markerssearch;
 
             var markerquery = from x in _context.Users.Include(u => u.Center).Include(u => u.Gender).Include(u => u.Role) select x;
@@ -72,7 +48,7 @@ namespace Markers_GPS_Coordiantes.Controllers
             {
                 markerquery = markerquery.Where(x => x.Loginname.Contains(markerssearch) || x.Firstname.Contains(markerssearch) || x.Lastname.Contains(markerssearch));
             }
-            return View(await markerquery.Where(b => b.CenterId == center.CenterId).OrderBy(r => r.Loginname).ToListAsync());
+            return View(await markerquery.AsNoTracking().ToListAsync());
 
 
         }
@@ -80,7 +56,17 @@ namespace Markers_GPS_Coordiantes.Controllers
         // GET: Users/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            
+            //  CHECK PERMISSIONS  -- ADD THIS CODE TO ALL YOUR PROTECTED ACTIONS
+            roleID = Convert.ToInt32(_sessionAccessor.HttpContext.Session.GetInt32("roleID"));
+            if (roleID <= 0)
+            {
+                return Unauthorized("You are not signed in.");          //  write better message
+            }
+            if (roleID != (int)RoleIDs.SuperAdmin)
+            {
+                return Unauthorized("You don't have permission to perform this operation.");  //  write better message
+            }
+            //  END OF SECURITY CHECK
             if (id == null)
             {
                 return NotFound();
@@ -102,8 +88,18 @@ namespace Markers_GPS_Coordiantes.Controllers
         // GET: Users/Create
         public IActionResult Create()
         {
-            
-        
+
+            //  CHECK PERMISSIONS  -- ADD THIS CODE TO ALL YOUR PROTECTED ACTIONS
+            roleID = Convert.ToInt32(_sessionAccessor.HttpContext.Session.GetInt32("roleID"));
+            if (roleID <= 0)
+            {
+                return Unauthorized("You are not signed in.");          //  write better message
+            }
+            if (roleID != (int)RoleIDs.SuperAdmin)
+            {
+                return Unauthorized("You don't have permission to perform this operation.");  //  write better message
+            }
+            //  END OF SECURITY CHECK
 
             ViewData["CenterId"] = new SelectList(_context.Center, "CenterId", "CenterName");
             ViewData["GenderId"] = new SelectList(_context.Gender, "GenderId", "GenderDescription");
@@ -118,6 +114,17 @@ namespace Markers_GPS_Coordiantes.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("UsersId,RoleId,CenterId,UsersToken,Loginname,Password,GenderId,Firstname,Lastname,EmailAddress,MobileNo,Telephone,Displayname,PostalAddress,PhysicalAddress,LastModifiedByUsersId,LastModifiedDate,CreatedByUsersId,CreateDate,IsDeleted")] Users users)
         {
+            //  CHECK PERMISSIONS  -- ADD THIS CODE TO ALL YOUR PROTECTED ACTIONS
+            roleID = Convert.ToInt32(_sessionAccessor.HttpContext.Session.GetInt32("roleID"));
+            if (roleID <= 0)
+            {
+                return Unauthorized("You are not signed in.");          //  write better message
+            }
+            if (roleID != (int)RoleIDs.SuperAdmin)
+            {
+                return Unauthorized("You don't have permission to perform this operation.");  //  write better message
+            }
+            //  END OF SECURITY CHECK
             if (ModelState.IsValid)
             {
                 _context.Add(users);
@@ -133,7 +140,17 @@ namespace Markers_GPS_Coordiantes.Controllers
     // GET: Users/Edit/5
     public async Task<IActionResult> Edit(int? id)
         {
-            
+            //  CHECK PERMISSIONS  -- ADD THIS CODE TO ALL YOUR PROTECTED ACTIONS
+            roleID = Convert.ToInt32(_sessionAccessor.HttpContext.Session.GetInt32("roleID"));
+            if (roleID <= 0)
+            {
+                return Unauthorized("You are not signed in.");          //  write better message
+            }
+            if (roleID != (int)RoleIDs.SuperAdmin)
+            {
+                return Unauthorized("You don't have permission to perform this operation.");  //  write better message
+            }
+            //  END OF SECURITY CHECK
             if (id == null)
             {
                 return NotFound();
@@ -157,6 +174,17 @@ namespace Markers_GPS_Coordiantes.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("UsersId,RoleId,CenterId,UsersToken,Loginname,Password,GenderId,Firstname,Lastname,EmailAddress,MobileNo,Telephone,Displayname,PostalAddress,PhysicalAddress,LastModifiedByUsersId,LastModifiedDate,CreatedByUsersId,CreateDate,IsDeleted")] Users users)
         {
+            //  CHECK PERMISSIONS  -- ADD THIS CODE TO ALL YOUR PROTECTED ACTIONS
+            roleID = Convert.ToInt32(_sessionAccessor.HttpContext.Session.GetInt32("roleID"));
+            if (roleID <= 0)
+            {
+                return Unauthorized("You are not signed in.");          //  write better message
+            }
+            if (roleID != (int)RoleIDs.SuperAdmin)
+            {
+                return Unauthorized("You don't have permission to perform this operation.");  //  write better message
+            }
+            //  END OF SECURITY CHECK
             if (id != users.UsersId)
             {
                 return NotFound();
@@ -181,6 +209,7 @@ namespace Markers_GPS_Coordiantes.Controllers
                     }
                 }
                 return RedirectToAction(nameof(Index));
+
             }
             ViewData["CenterId"] = new SelectList(_context.Center, "CenterId", "CenterName", users.CenterId);
             ViewData["GenderId"] = new SelectList(_context.Gender, "GenderId", "GenderDescription", users.GenderId);
@@ -191,7 +220,17 @@ namespace Markers_GPS_Coordiantes.Controllers
         // GET: Users/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-           
+            //  CHECK PERMISSIONS  -- ADD THIS CODE TO ALL YOUR PROTECTED ACTIONS
+            roleID = Convert.ToInt32(_sessionAccessor.HttpContext.Session.GetInt32("roleID"));
+            if (roleID <= 0)
+            {
+                return Unauthorized("You are not signed in.");          //  write better message
+            }
+            if (roleID != (int)RoleIDs.SuperAdmin)
+            {
+                return Unauthorized("You don't have permission to perform this operation.");  //  write better message
+            }
+            //  END OF SECURITY CHECK
             if (id == null)
             {
                 return NotFound();
@@ -215,6 +254,17 @@ namespace Markers_GPS_Coordiantes.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            //  CHECK PERMISSIONS  -- ADD THIS CODE TO ALL YOUR PROTECTED ACTIONS
+            roleID = Convert.ToInt32(_sessionAccessor.HttpContext.Session.GetInt32("roleID"));
+            if (roleID <= 0)
+            {
+                return Unauthorized("You are not signed in.");          //  write better message
+            }
+            if (roleID != (int)RoleIDs.SuperAdmin)
+            {
+                return Unauthorized("You don't have permission to perform this operation.");  //  write better message
+            }
+            //  END OF SECURITY CHECK
             var users = await _context.Users.FindAsync(id);
             _context.Users.Remove(users);
             await _context.SaveChangesAsync();
@@ -223,6 +273,7 @@ namespace Markers_GPS_Coordiantes.Controllers
 
         private bool UsersExists(int id)
         {
+
             return _context.Users.Any(e => e.UsersId == id);
         }
     }
